Rename core.py to rich_click.py for more intuitive settings naming
